abstract all prefetch functions
